requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.11
django-cors-headers==4.4.0
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==2.2.0
django-richtextfield==1.6.2
djangorestframework==3.15.2
et-xmlfile==1.1.0
gunicorn==23.0.0
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pypng==0.20220715.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
